requirements.txt
streamlit
langchain
langchain-groq
pypdf
langchain-huggingface
langchain-community
python-dotenv
faiss-cpu
sentence-transformers
